Major refactor
Not really the way of working
