+++ config/colorTheme.js
@@ -1,0 +1,11 @@
+module.exports = {
+  input: 'grey',
+  verbose: 'cyan',
+  prompt: 'grey',
+  info: 'green',
+  data: 'grey',
+  help: 'cyan',
+  debug: 'cyan',
+  warn: 'orange',
+  error: 'red'
+};
