+++ public/js/app.js
@@ -1,0 +1,12 @@
+"use strict";
+
+/**
+ * Initializes RxVisualSedimentation
+ */
+var init = function () {
+  initButtons();
+  initClockObversable();
+  initStateObservable();
+  initEnvironment();
+  initWebsocket();
+};
