+++ public/js/websocket.js
@@ -1,26 +1,22 @@
-$(function () {
-  var content = $('#content');
-  var input = $('#input');
-  var status = $('#status');
-  
+var initWebsocket = function () {
   window.WebSocket = window.WebSocket || window.MozWebSocket;
-  
+
   var connection = new WebSocket('ws://127.0.0.1:3000/', 'twitter-protocol');
-  
+
   if (!window.WebSocket) {
     console.log("Your browser does not support websockets");
     return;
   } 
-  
+
   connection.onopen = function () {
     console.log("Connection succesfully opened");
   }
-  
+
   connection.onerror = function (error) { 
     console.log("Error occurred: ");
     console.log(error.toString());
   }
-  
+
   connection.onmessage = function (message) {
     try {
       var tweet = JSON.parse(message.data);
@@ -29,17 +25,17 @@ $(function () {
       return;
     }
     console.log(tweet.created_at + " - " + tweet.user.name);
-  }
+  }  
+}
   
-  input.keydown(function(e) {
-    if (e.keyCode === 13) {
-      var msg = $(this).val();
-      if (!msg) {
-        return;
-      }
-      // send the message as an ordinary text
-      connection.send(msg.toString());
-      $(this).val('');
-    }
-  });  
-});
+//  input.keydown(function(e) {
+//    if (e.keyCode === 13) {
+//      var msg = $(this).val();
+//      if (!msg) {
+//        return;
+//      }
+//      // send the message as an ordinary text
+//      connection.send(msg.toString());
+//      $(this).val('');
+//    }
+//  });
